move helper functions to dialect
